# frontend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Trader API",
    "description": "Русскоязычная документация API сервиса AI Trader на FastAPI.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/market/instruments": {
      "get": {
        "tags": ["market"],
        "summary": "Список инструментов",
        "description": "Возвращает список рыночных инструментов.",
        "operationId": "market_instruments_api_v1_market_instruments_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "default": 0, "title": "Offset" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/recommendations": {
      "get": {
        "tags": ["market"],
        "summary": "Список рекомендаций",
        "description": "Возвращает рекомендации по инструментам.",
        "operationId": "market_recommendations_api_v1_market_recommendations_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "default": 0, "title": "Offset" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/recommendations/{figi}": {
      "get": {
        "tags": ["market"],
        "summary": "Рекомендация по FIGI",
        "description": "Последняя рекомендация по инструменту (тот же DTO, что в списке).",
        "operationId": "market_recommendation_by_figi_api_v1_market_recommendations__figi__get",
        "parameters": [
          {
            "name": "figi",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Figi" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/stock/{figi}": {
      "get": {
        "tags": ["market"],
        "summary": "Карточка инструмента по FIGI",
        "description": "Возвращает детальную карточку инструмента.",
        "operationId": "market_stock_api_v1_market_stock__figi__get",
        "parameters": [
          {
            "name": "figi",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Figi" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/stock/{figi}/candles": {
      "get": {
        "tags": ["market"],
        "summary": "Свечи инструмента по FIGI",
        "description": "Возвращает историю свечей для инструмента.",
        "operationId": "market_stock_candles_api_v1_market_stock__figi__candles_get",
        "parameters": [
          {
            "name": "figi",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Figi" }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 0, "default": 0, "title": "Offset" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3650,
              "minimum": 1,
              "default": 365,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/stock/{figi}/analyst-signals": {
      "get": {
        "tags": ["market"],
        "summary": "Сигналы аналитиков по FIGI",
        "description": "Сигналы из БД (синхронизация scheduler signals_update).",
        "operationId": "market_stock_analyst_signals_api_v1_market_stock__figi__analyst_signals_get",
        "parameters": [
          {
            "name": "figi",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Figi" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/stock/{figi}/weekly-forecast": {
      "get": {
        "tags": ["market"],
        "summary": "Weekly LSTM: прогноз (из БД или refresh)",
        "description": "Сохранённый weekly-прогноз по рекомендации; при refresh — повторный инференс.",
        "operationId": "market_stock_weekly_forecast_api_v1_market_stock__figi__weekly_forecast_get",
        "parameters": [
          {
            "name": "figi",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Figi" }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "title": "Refresh" },
            "description": "Пересчитать модель и записать в БД"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessEnvelope_dict_str__object__" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessEnvelope_dict_str__object__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Флаг успешного ответа",
            "default": true
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "Обязательное поле: полезная нагрузка ответа"
          }
        },
        "type": "object",
        "required": ["data"],
        "title": "SuccessEnvelope[dict[str, object]]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}